Fix last remaining tests
